# contracts/host/openapi.json
{
  "components": {
    "schemas": {
...
      "InventorySearchItem": {
        "additionalProperties": false,
        "properties": {
          "contextIds": {
            "items": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            },
            "type": "array"
          },
          "createdAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?Z$",
            "type": "string"
          },
          "currentRevisionId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "displayName": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "itemId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "itemType": {
            "anyOf": [
              {
                "enum": [
                  "game"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "analysis"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "source"
                ],
                "type": "string"
              }
            ]
          },
          "languageTag": {
            "maxLength": 35,
            "minLength": 1,
            "type": "string"
          },
          "originKind": {
            "anyOf": [
              {
                "enum": [
                  "manual"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "structured_import"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "playout"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "live_observed"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "live_played"
                ],
                "type": "string"
              }
            ]
          },
          "rootAnchorId": {
            "pattern": "^[1-9]\\d{0,15}$",
            "type": "string"
          },
          "summary": {
            "maxLength": 2000,
            "type": "string"
          },
          "updatedAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?Z$",
            "type": "string"
          }
        },
        "required": [
          "itemId",
          "currentRevisionId",
          "rootAnchorId",
          "itemType",
          "originKind",
          "displayName",
          "languageTag",
          "contextIds",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "correlationId": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "parameters": {
            "additionalProperties": false,
            "properties": {
              "currentRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "expectedRevision": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              }
            },
            "type": "object"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "retryable",
          "parameters"
        ],
        "type": "object"
      },
...
      "SearchInventoryResult": {
        "additionalProperties": false,
        "properties": {
          "dataRevision": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InventorySearchItem"
            },
            "type": "array"
          },
          "nextCursor": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "items",
          "dataRevision"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "hostBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Plysmith Application Host",
    "version": "0.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/inventory": {
      "get": {
        "operationId": "SearchInventory",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "contextId",
            "required": false,
            "schema": {
              "pattern": "^[1-9]\\d{0,15}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "pattern": "^(?:[1-9]|[1-9]\\d|100)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchInventoryResult"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "A safe, stable Plysmith problem."
          }
        }
      }
    },
...
  },
  "security": [
    {
      "hostBearer": []
    }
  ]
}